# contracts/openapi/payflow-complete.yaml
openapi: 3.1.0
info:
  title: PayFlow API - Advanced Fintech Benchmark
  version: 1.0.0
  description: |
    Comprehensive API for fintech microservices benchmarking.
    Implements advanced patterns: Saga orchestration, CQRS, Event Sourcing,
    idempotency, and distributed transaction management.
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.payflow.example.com
    description: Production server
paths:
...
  /v1/accounts:
    post:
      summary: Create new account
      tags: [Accounts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccountRequest"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountResponse"
        "400":
          description: Invalid request
        "409":
          description: Account already exists
    
    get:
      summary: List accounts
      tags: [Accounts]
      parameters:
        - name: user_id
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, suspended]
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Accounts list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountListResponse"
  /v1/accounts/{account_id}:
    get:
      summary: Get account details
      tags: [Accounts]
      parameters:
        - name: account_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Account details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountResponse"
        "404":
          description: Account not found
    
    patch:
      summary: Update account
      tags: [Accounts]
      parameters:
        - name: account_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAccountRequest"
      responses:
        "200":
          description: Account updated
        "404":
          description: Account not found
        "400":
          description: Invalid update
...
components:
  schemas:
...
    CreateAccountRequest:
      type: object
      required:
        - user_id
        - currency
        - account_type
      properties:
        user_id:
          type: string
        currency:
          type: string
          minLength: 3
          maxLength: 3
          example: "USD"
        account_type:
          type: string
          enum: [checking, savings, business]
        initial_balance:
          type: number
          minimum: 0
          default: 0
        metadata:
          type: object
          additionalProperties: true
    AccountResponse:
      type: object
      required:
        - account_id
        - user_id
        - account_number
        - currency
        - status
        - created_at
      properties:
        account_id:
          type: string
        user_id:
          type: string
        account_number:
          type: string
        currency:
          type: string
        account_type:
          type: string
        status:
          type: string
          enum: [active, inactive, suspended]
        balance:
          type: number
        available_balance:
          type: number
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    UpdateAccountRequest:
      type: object
      properties:
        status:
          type: string
          enum: [active, inactive, suspended]
        metadata:
          type: object
          additionalProperties: true
...
    AccountListResponse:
      type: object
      required:
        - accounts
        - pagination
      properties:
        accounts:
          type: array
          items:
            $ref: "#/components/schemas/AccountResponse"
        pagination:
          $ref: "#/components/schemas/Pagination"
    
    # Transactions
...
    Pagination:
      type: object
      required:
        - page
        - limit
        - total
        - pages
      properties:
        page:
          type: integer
          minimum: 1
        limit:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
        pages:
          type: integer
          minimum: 0
        has_next:
          type: boolean
        has_prev:
          type: boolean
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
...